Implement getPlayerSelection function
The function takes no parameters and returns the user's selection in
the form of a string. The user is prompted and then the answer is
normalised. The answer is then validated in order to ensure the user is
following the rules and ensure correct comparisons further on in the
code. If the selection of the user is invalid, the user will be alerted.